test for create_slideshow_manifest in publish
